refactor(formatter): move formatters to utils specific folder
