+++ packages/sp-proxy/src/interface-adapters/utils/formatters/GenerateMetadataFormatter.spec.ts
@@ -1,4 +1,4 @@
-import { GenerateMetadataFormatter } from '@sp-proxy/interface-adapters/utils/GenerateMetadataFormatter'
+import { GenerateMetadataFormatter } from '@sp-proxy/interface-adapters/utils/formatters/GenerateMetadataFormatter'
 
 describe('GenerateMetadataTransformer', () => {
   it('should call replace with correct values', async () => {
