+++ packages/sp-proxy/src/interface-adapters/utils/formatters/GenerateMetadataFormatter.ts
